Changing test system to include user defined tests.
Currently tests will will not work just checking out the repo, as the test
class is referring to a user directory that is currently being ignored.
This is a temporary solution to allow separation of the system core for the
mbed project specific UDFs and tests. This will be resolved later with use
of git sub-modules.